refactored navigation into one componen, removed double code
