+++ src/components/navigation/Navigation.styled.ts
@@ -1,5 +1,13 @@
 import styled from 'styled-components';
 
+export const Wrapper = styled.nav<{ isExtraRoute: boolean }>`
+  display: flex;
+  justify-content: ${({ isExtraRoute }) =>
+    !isExtraRoute ? `space-between` : `none`};
+  padding: ${({ theme }) => theme.space[1]}px;
+  background-color: ${({ theme }) => theme.colors.accentBackgroundColor};
+`;
+
 export const NavigationListStyled = styled.ul`
   background-color: ${p => p.theme.colors.accentBackgroundColor};
   display: flex;
@@ -8,10 +16,9 @@ export const NavigationListStyled = styled.ul`
   text-align: center;
 `;
 
-export const Wrapper = styled.nav<{ isExtraRoute: boolean }>`
+export const PageName = styled.p`
   display: flex;
-  justify-content: ${({ isExtraRoute }) =>
-    !isExtraRoute ? `space-between` : `none`};
-  padding: ${({ theme }) => theme.space[1]}px;
-  background-color: ${({ theme }) => theme.colors.accentBackgroundColor};
+  align-items: center;
+  margin-left: ${({ theme }) => theme.space[3]}px;
+  color: ${({ theme }) => theme.colors.lightText};
 `;
