+++ src/components/navigation/extraMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ExtraMenu';
