--- src/components/navigation/navigationCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationCard';
