--- src/components/navigation/navigationMain/NavigationMain.styled.ts
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const MenuHolder = styled.div`
-  display: flex;
-`;
