--- src/components/navigation/navigationMain/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationMain';
