+++ src/constants/paths.ts
@@ -2,6 +2,8 @@ export const paths: IPaths = {
   home: '/',
   ingredients: '/ingredients',
   cocktails: '/cocktails',
+  cocktailCard: '/cocktails/:cocktailId',
+  ingredientCard: 'ingredients/:ingredientId',
   settings: '/settings',
   newCocktail: '/cocktails/new',
   newIngredient: '/ingredients/new',
